requirements.txt
asgiref==3.5.0
Django==3.2.13
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
pkg-resources==0.0.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.2.0
